requirements.txt
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.7
cycler==0.11.0
dnspython==2.1.0
fake-rpi @ git+https://github.com/sn4k3/FakeRPi@0f30d320d5f715d8a4fb94e7105448508586ae94
feedparser==6.0.8
fonttools==4.28.3
gpio==0.3.0
idna==3.3
kiwisolver==1.3.2
LinkChecker @ git+https://github.com/linkchecker/linkchecker.git@900586dc0117ba23b68309266bc0498d83ac1f7e
matplotlib==3.5.0
mpmath==1.2.1
np==1.0.2
numpy==1.21.4
packaging==21.3
Pillow==8.4.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-dotenv==0.19.2
pyxdg==0.27
requests==2.26.0
setuptools-scm==6.3.2
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.3.1
sympy==1.9
tomli==1.2.2
urllib3==1.26.7
